Cleaned up settings code
