+++ src/program/devgui/homemenu/HomeMenuSettings.h
@@ -3,7 +3,7 @@
 #include "imgui.h"
 #include "imgui_internal.h"
 
-#include "devgui/savedata/DevGuiSettings.h"
+#include "devgui/settings/DevGuiSettings.h"
 #include "devgui/homemenu/HomeMenuBase.h"
 
 class HomeMenuSettings : public HomeMenuBase {
