+++ src/program/devgui/settings/DevGuiSettings.cpp
@@ -1,39 +1,34 @@
 #include "devgui/DevGuiManager.h"
 
-#include "devgui/savedata/DevGuiSettings.h"
+#include "devgui/settings/DevGuiSettings.h"
 
 // Settings are added and created here
 DevGuiSettings::DevGuiSettings(DevGuiManager* parent, sead::Heap* heap)
 {
+    mSettingsHeap = heap;
+
     // Max of 0x20 (32) settings, should never go beyond this but if you do for some reason, increase this number!
     mSettings.allocBuffer(0x20, heap);
 
-    DevGuiSettingsEntry* autosave = new (heap) DevGuiSettingsEntry(true, true, "Autosave");
-    mSettings.pushBack(autosave);
-
-    DevGuiSettingsEntry* hud = new (heap) DevGuiSettingsEntry(true, true, "Display HUD");
-    mSettings.pushBack(hud);
-
-    DevGuiSettingsEntry* music = new (heap) DevGuiSettingsEntry(true, true, "Play Music");
-    mSettings.pushBack(music);
-
-    DevGuiSettingsEntry* checkpoint = new (heap) DevGuiSettingsEntry(false, true, "Always Allow Checkpoints");
-    mSettings.pushBack(checkpoint);
-
-    DevGuiSettingsEntry* noclip = new (heap) DevGuiSettingsEntry(false, false, "Noclip");
-    mSettings.pushBack(noclip);
-
-    DevGuiSettingsEntry* gmr = new (heap) DevGuiSettingsEntry(false, false, "Grey Moon Refresh");
-    mSettings.pushBack(gmr);
-
-    DevGuiSettingsEntry* buttonmotionroll = new (heap) DevGuiSettingsEntry(false, false, "Button Motion Roll");
-    mSettings.pushBack(buttonmotionroll);
-
-    DevGuiSettingsEntry* nodamage = new (heap) DevGuiSettingsEntry(false, false, "No Damage");
-    mSettings.pushBack(nodamage);
+    registerNewSetting(false, false, "Noclip");
+    registerNewSetting(false, true, "Button Motion Roll");
+    registerNewSetting(false, true, "Grey Moon Refresh");
+    registerNewSetting(false, true, "No Damage");
+    registerNewSetting(false, true, "Always Allow Checkpoints");
+    registerNewSetting(true, true, "Autosave");
+    registerNewSetting(true, true, "Display HUD");
+    registerNewSetting(true, true, "Play Music");
 
     Logger::log("Created %i settings\n", mSettings.size());
 }
+
+void DevGuiSettings::registerNewSetting(bool isEnabledByDefault, bool isSave, const char* settingName)
+{
+    DevGuiSettingsEntry* newSet = new (mSettingsHeap) DevGuiSettingsEntry(isEnabledByDefault, isSave, settingName);
+    mSettings.pushBack(newSet);
+}
+
+
 
 // Everything below this point is general helper functions, likely not what you are looking for!
