+++ src/program/devgui/settings/HooksSettings.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "game/GameData/GameDataHolder.h"
+#include "game/GameData/GameDataHolderWriter.h"
+#include "game/GameData/PlayerHitPointData.h"
+#include "game/StageScene/StageScene.h"
+
+#include "devgui/DevGuiManager.h"
+
+#include "lib.hpp"
+
+void exlSetupSettingsHooks();
