+++ src/program/main.cpp
@@ -65,6 +65,7 @@
 #include "GetterUtil.h"
 #include "devgui/DevGuiManager.h"
 #include "devgui/DevGuiPrimitive.h"
+#include "devgui/settings/HooksSettings.h"
 
 #include <typeinfo>
 
@@ -86,28 +87,6 @@ sead::TextWriter *gTextWriter;
 void drawLunaKit() {
     DevGuiManager::instance()->updateDisplay();
 }
-
-HOOK_DEFINE_TRAMPOLINE(ControlHook) {
-    static void Callback(StageScene *scene) {
-        DevGuiManager::instance()->updateNoclip();
-
-        DevGuiSettings* set = DevGuiManager::instance()->getSettings();
-
-        if(!set->getStateByName("Display HUD") && scene->mSceneLayout->isWait()) {
-            scene->mSceneLayout->end();
-            MapMini* compass = scene->mSceneLayout->mMapMiniLyt;
-            if (compass->mIsAlive) compass->end();
-        }
-
-        if(!set->getStateByName("Play Music")) {
-            if (al::isPlayingBgm(scene)) {
-                al::stopAllBgm(scene, 0);
-            }
-        }
-
-        Orig(scene);
-    }
-};
 
 HOOK_DEFINE_TRAMPOLINE(CreateFileDeviceMgr) {
     static void Callback(sead::FileDeviceMgr *thisPtr) {
@@ -252,51 +231,6 @@ HOOK_DEFINE_TRAMPOLINE(DrawLunaPrimitives) {
     }
 };
 
-HOOK_DEFINE_TRAMPOLINE(SaveHook) {
-    static bool Callback(StageScene* scene) {
-        if (DevGuiManager::instance()->getSettings()->getStateByName("Autosave"))
-            return Orig(scene);
-        else
-            return false;
-    }
-};
-
-HOOK_DEFINE_TRAMPOLINE(CheckpointWarpHook) {
-    static bool Callback(void* thisPtr) {
-        if (DevGuiManager::instance()->getSettings()->getStateByName("Always Allow Checkpoints"))
-            return true;
-        else
-            return Orig(thisPtr);
-    }
-};
-HOOK_DEFINE_TRAMPOLINE(GreyShineRefreshHook) {
-    static bool Callback(GameDataHolderWriter writer, ShineInfo const* shineInfo) {
-        if (DevGuiManager::instance()->getSettings()->getStateByName("Grey Moon Refresh"))
-            return false;
-        return Orig(writer, shineInfo);
-
-        
-    }
-};
-HOOK_DEFINE_TRAMPOLINE(ButtonMotionRollHook) {
-    static bool Callback(void* thisPtr) {
-        if (DevGuiManager::instance()->getSettings()->getStateByName("Button Motion Roll")){
-            return true;
-            }
-        else
-            return Orig(thisPtr);
-    }
-};
-
-HOOK_DEFINE_TRAMPOLINE(NoDamageHook){
-    static void Callback(PlayerHitPointData* hitPointData) {
-        if (!DevGuiManager::instance()->getSettings()->getStateByName("No Damage"))
-            return Orig(hitPointData);
-
-    }
-};
-
-
 extern "C" void exl_main(void *x0, void *x1) {
     /* Setup hooking enviroment. */
     // envSetOwnProcessHandle(exl::util::proc_handle::Get());
@@ -321,15 +255,9 @@ extern "C" void exl_main(void *x0, void *x1) {
     // Debug Text Writer Drawing
     DrawLunaPrimitives::InstallAtOffset(0x50F1D8);
 
-    // General Hooks
-    ControlHook::InstallAtSymbol("_ZN10StageScene7controlEv");
-
     // DevGui cheats
-    SaveHook::InstallAtSymbol("_ZNK10StageScene12isEnableSaveEv");
-    CheckpointWarpHook::InstallAtOffset(0x1F2998);
-    GreyShineRefreshHook::InstallAtSymbol("_ZN16GameDataFunction10isGotShineE22GameDataHolderAccessorPK9ShineInfo");
-    ButtonMotionRollHook::InstallAtSymbol("_ZNK23PlayerJudgeStartRolling21isTriggerRestartSwingEv");
-    NoDamageHook::InstallAtSymbol("_ZN16GameDataFunction12damagePlayerE20GameDataHolderWriter");
+    exlSetupSettingsHooks(); // Located in devgui/settings/HooksSettings
+
     // ImGui Hooks
 #if IMGUI_ENABLED
     nvnImGui::InstallHooks();
